Move сalculations from Card Component to Card Model
