+++ src/components/card.js
@@ -1,26 +1,3 @@
-import {castTimeFormat, getCorrectTime} from "../helpers";
-
-const countDuration = (dateFrom, dateTo) => {
-  let durationString = `00M`;
-  const durationInMinutes = Math.floor((dateTo - dateFrom) / (1000 * 60));
-
-  if (durationInMinutes > 0) {
-    durationString = `${castTimeFormat(durationInMinutes % 60)}M`;
-  }
-
-  if (durationInMinutes >= 60) {
-    const durationInHours = Math.floor(durationInMinutes / 60);
-    durationString = `${castTimeFormat(durationInHours % 24)}H ${durationString}`;
-
-    if (durationInHours >= 24) {
-      const durationInDays = Math.floor(durationInHours / 24);
-      durationString = `${castTimeFormat(durationInDays)}D ${durationString}`;
-    }
-  }
-
-  return durationString;
-};
-
 const createOffersTemplate = (offers) => {
   return Array.from(offers)
     .map((offer) => {
@@ -35,10 +12,7 @@ const createOffersTemplate = (offers) => {
 };
 
 export const createCardTemplate = (card) => {
-  const {type, icon, city, dateFrom, dateTo, price, offers} = card;
-  const correctDateFrom = getCorrectTime(dateFrom);
-  const correctDateTo = getCorrectTime(dateTo);
-  const duration = countDuration(dateFrom, dateTo);
+  const {type, icon, city, correctDateFrom, correctDateTo, price, offers, duration} = card;
   const offersTemplate = createOffersTemplate(offers);
 
   return (
